+++ reportcard-server/src/main/java/io/github/ericdriggs/reportcard/controller/browse/BrowseHtmlHelper.java
@@ -7,6 +7,7 @@ import io.github.ericdriggs.reportcard.gen.db.tables.pojos.*;
 import io.github.ericdriggs.reportcard.model.StageTestResultPojo;
 import io.github.ericdriggs.reportcard.model.branch.BranchJobLatestRunMap;
 import io.github.ericdriggs.reportcard.model.branch.RunStorageTestResult;
+import io.github.ericdriggs.reportcard.persist.StorageType;
 import io.github.ericdriggs.reportcard.util.NumberStringUtil;
 import io.github.ericdriggs.reportcard.util.PrettyPrintUtil;
 import io.github.ericdriggs.reportcard.util.StringMapUtil;
@@ -410,12 +411,31 @@ public class BrowseHtmlHelper {
         }
         StringBuilder sb = new StringBuilder();
         for (StoragePojo storage : storages) {
-            final String reportLink = reportLinkBase
-                    .replace("{reportName}", storage.getLabel())
-                    .replace("{reportUrl}", getStorageKey(storage));
-            sb.append(reportLink + ls);
+            final String displayName = getDisplayName(storage);
+            if (storage.getStorageType() != null
+                && storage.getStorageType() == StorageType.TAR_GZ.getStorageTypeId()) {
+                // Download link for tar.gz files with tooltip
+                final String downloadLink = downloadLinkBase
+                        .replace("{downloadName}", displayName)
+                        .replace("{downloadUrl}", getStorageKey(storage))
+                        .replace("{tooltip}", "Download " + storage.getLabel());
+                sb.append(downloadLink + ls);
+            } else {
+                // Existing browse link for HTML/other files
+                final String reportLink = reportLinkBase
+                        .replace("{reportName}", displayName)
+                        .replace("{reportUrl}", getStorageKey(storage));
+                sb.append(reportLink + ls);
+            }
         }
         return sb.toString();
+    }
+
+    protected static String getDisplayName(StoragePojo storage) {
+        if (storage.getIndexFile() == null) {
+            return storage.getLabel();
+        }
+        return storage.getIndexFile();
     }
 
     protected final static String stageViewMain =
@@ -464,10 +484,15 @@ public class BrowseHtmlHelper {
             </fieldset>
             """;
 
-    protected final static String reportLinkBase = "<a class=\"info report-link\" href=\"{reportUrl}\">" +
+    protected final static String reportLinkBase = "<a class=\"info report-link\" style=\"white-space: nowrap\" href=\"{reportUrl}\">" +
                                                    "<img alt=\"{reportName}\" class=\"report-img\" src=\"/image/report-simple.svg\">" +
                                                    "{reportName}" +
                                                    "</a>";
+
+    protected final static String downloadLinkBase = "<a class=\"info report-link\" style=\"white-space: nowrap\" href=\"{downloadUrl}\" title=\"{tooltip}\">" +
+                                                     "<img alt=\"{downloadName}\" class=\"report-img\" src=\"/image/report-simple.svg\">" +
+                                                     "{downloadName}" +
+                                                     "</a>";
 
     //******************** job stages ********************//
 
@@ -748,7 +773,7 @@ public class BrowseHtmlHelper {
                 &nbsp;&nbsp;&nbsp;
                 <a href="/swagger-ui/index.html">swagger</a>&nbsp;&nbsp;&nbsp;
                 <a href="https://github.com/ericdriggs/reportcard">source</a>&nbsp;&nbsp;&nbsp;
-                <span style="color:white">ver: 0.1.27</span>
+                <span style="color:white">ver: 0.1.28</span>
               </span>
             </header>
             <nav aria-label="breadcrumb">